agent/requirements.txt
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-generativeai==0.8.3
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
schedule==1.2.1
